+++ src/main/java/kitchenpos/products/exception/ProductErrorCode.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.exception.message;
+package kitchenpos.products.exception;
 
 public enum ProductErrorCode {
